update security, and methods in mvc

+++ itSpaceQuiz-mvc/src/main/java/com/example/itspacequizmvc/service/QuestionService.java
@@ -2,25 +2,23 @@ package com.example.itspacequizmvc.service;
 
 import com.example.itspacequizcommon.entity.Question;
 import com.example.itspacequizcommon.entity.Quiz;
-import com.example.itspacequizcommon.repository.QuestionOptionRepository;
 import com.example.itspacequizcommon.repository.QuestionRepository;
+import com.example.itspacequizmvc.exception.NotFoundException;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class QuestionService {
     private final QuestionRepository questionRepository;
-    private final QuestionOptionRepository questionOptionRepository;
 
-    public List<Question> getAllQuestions() {
-        return questionRepository.findAll();
-
-    }
 
     public Question save(Question question) {
 
@@ -30,11 +28,20 @@ public class QuestionService {
 
 
     public void deleteById(int id) {
+        Optional<Question> question = questionRepository.findById(id);
+        if (question.isEmpty()) {
+            throw new NotFoundException("Question not found with id: " + id);
+        }
         questionRepository.deleteById(id);
+        log.info("Question with id {} deleted", id);
+
     }
 
 
     public Question findById(int id) {
+        if (!questionRepository.existsById(id)){
+            throw new NotFoundException("Question not found for provided id.");
+        }
         return questionRepository.getById(id);
     }
 
@@ -43,15 +50,12 @@ public class QuestionService {
 
         return questionRepository.findAllByQuiz(quiz);
     }
+
     public Page<Question> findAll(Pageable pageable) {
 
         return questionRepository.findAll(pageable);
     }
 
-    public Question findByQuiz(Quiz quiz){
-        return questionRepository.findByQuiz(quiz);
-    }
 
-
-    }
+}
